added serializer for todo and used it in the todo controller

+++ app/controllers/todos_controller.rb
@@ -1,16 +1,26 @@
 class TodosController < ApplicationController
   def index
-    todos = Todo.all
+
+    todos = ActiveModel::ArraySerializer.new(Todo.all)
     render json: todos
   end
 
   def create
-    todo = Todo.create(title: params[:todo][:title])
-    render json: todo
+    todo = Todo.new(title: params[:todo][:title])
+    if todo.save
+      render json: todo
+    else
+      render json: {errors: todo.errors.full_messagese}
+    end
   end
 
   def destroy
     todo = Todo.find(params[:id])
-    todo.destroy
+    if todo.present?
+      todo.destroy
+      render json: {id: todo.id}
+    else
+      render json: {error: "No todo found with that id"}
+    end
   end
 end
